requirements.txt
slack_sdk
notion-client
python-dotenv
apscheduler
